run password authentication tests using a TTY
